requirements-prod.txt
# Requirements optimizados para producción
# Solo dependencias esenciales para runtime

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning (versión CPU para reducir tamaño)
tensorflow-cpu==2.15.0
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
Pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Data processing (solo para entrenamiento)
pandas==2.1.4

# Excluidas para producción (no necesarias en runtime):
# matplotlib (solo para visualizaciones)
# seaborn (solo para visualizaciones) 
# requests (solo para testing)
